third_party/blink/web_tests/external/wpt/shadow-dom/event-post-dispatch-expected.txt

This is a testharness.js-based test.
[FAIL] Event properties post dispatch when target get moved into the shadow tree by event listener
  assert_equals: Target should not have been cleared expected Element node <trala></trala> but got null
Harness: the test ran to completion.
